rector

+++ src/Controller/Backend/BesoinController.php
@@ -15,7 +15,6 @@ use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Security\Http\Attribute\IsGranted;
 
-#[Route(path: '/backend/besoin')]
 #[IsGranted('ROLE_VOLONTARIAT')]
 class BesoinController extends AbstractController
 {
@@ -23,15 +22,16 @@ class BesoinController extends AbstractController
 
     public function __construct(
         private readonly BesoinRepository $besoinRepository,
-        private readonly MessageBusInterface $dispatcher,
+        private readonly MessageBusInterface $messageBus,
     ) {}
 
-    #[Route(path: '/index', name: 'volontariat_backend_besoin', methods: ['GET'])]
+    #[Route(path: '/backend/besoin/index', name: 'volontariat_backend_besoin', methods: ['GET'])]
     public function index(): Response
     {
-        if (($hasAssociation = $this->hasAssociation()) !== null) {
+        if (($hasAssociation = $this->hasAssociation()) instanceof Response) {
             return $hasAssociation;
         }
+
         $annonces = $this->besoinRepository->findByAssociation($this->association);
 
         return $this->render(
@@ -43,12 +43,13 @@ class BesoinController extends AbstractController
         );
     }
 
-    #[Route(path: '/new', name: 'volontariat_backend_besoin_new', methods: ['GET', 'POST'])]
+    #[Route(path: '/backend/besoin/new', name: 'volontariat_backend_besoin_new', methods: ['GET', 'POST'])]
     public function new(Request $request): Response
     {
-        if (($hasAssociation = $this->hasAssociation()) !== null) {
+        if (($hasAssociation = $this->hasAssociation()) instanceof Response) {
             return $hasAssociation;
         }
+
         $besoin = new Besoin();
         $besoin->setAssociation($this->association);
         if (!$this->association->valider) {
@@ -56,6 +57,7 @@ class BesoinController extends AbstractController
 
             return $this->redirectToRoute('volontariat_dashboard');
         }
+
         $form = $this->createForm(BesoinType::class, $besoin);
 
         $form->handleRequest($request);
@@ -65,7 +67,7 @@ class BesoinController extends AbstractController
             $this->besoinRepository->persist($besoin);
             $this->besoinRepository->flush();
             $this->addFlash('success', 'Le besoin a bien été ajouté');
-            $this->dispatcher->dispatch(new BesoinCreated($besoin->getId()));
+            $this->messageBus->dispatch(new BesoinCreated($besoin->getId()));
 
             return $this->redirectToRoute('volontariat_backend_besoin');
         }
@@ -79,7 +81,7 @@ class BesoinController extends AbstractController
         );
     }
 
-    #[Route(path: '/{uuid}/edit', name: 'volontariat_backend_besoin_edit')]
+    #[Route(path: '/backend/besoin/{uuid}/edit', name: 'volontariat_backend_besoin_edit')]
     #[IsGranted('edit', subject: 'besoin')]
     public function edit(Request $request,#[MapEntity(expr: 'repository.findOneByUuid(uuid)')]  Besoin $besoin): Response
     {
@@ -104,18 +106,17 @@ class BesoinController extends AbstractController
         );
     }
 
-    #[Route(path: '/delete', name: 'volontariat_backend_besoin_delete', methods: ['POST'])]
+    #[Route(path: '/backend/besoin/delete', name: 'volontariat_backend_besoin_delete', methods: ['POST'])]
     public function delete(Request $request): RedirectResponse
     {
         if ($this->isCsrfTokenValid('deletedd', $request->request->get('_token'))) {
             $uid = $request->request->get('id');
             $besoin = $this->besoinRepository->findOneBy(['uuid' => $uid]);
-            if ($besoin) {
-                if ($this->isGranted('edit', $besoin)) {
-                    $this->besoinRepository->remove($besoin);
-                    $this->besoinRepository->flush();
-                }
+            if ($besoin && $this->isGranted('edit', $besoin)) {
+                $this->besoinRepository->remove($besoin);
+                $this->besoinRepository->flush();
             }
+
             $this->addFlash('success', 'Le besoin a bien été supprimé');
         }
